Refactor benches.
